Revert "Merge pull request #270 from duyquangnguyenhac/gemstash-milestone1-s3"
This reverts commit 872ccae8e3abb65a226e3d33715865c2fa6b1beb, reversing
changes made to 58437b71f5d6785f4bc1ebe3c75042c245ae4298.
