--- fixtures/vcr_cassettes/Gemstash_S3/_property_/with_a_single_key/returns_true_for_a_key_pointing_to_explicit_nil.yml
@@ -1,223 +1,0 @@
----
-http_interactions:
-- request:
-    method: put
-    uri: https://<REDACTED_BUCKET_NAME>.s3.<REDACTED_REGION_NAME>.amazonaws.com/gemstash/s3_storage/TEST_S3_SPEC_FOLDER/private/gems/existing-f4e0ac58eb46d88efc451c164db3b837/file
-    body:
-      encoding: UTF-8
-      string: content
-    headers:
-      Content-Type:
-      - ''
-      Accept-Encoding:
-      - ''
-      User-Agent:
-      - aws-sdk-ruby3/3.104.3 ruby/2.6.3 x86_64-darwin19 aws-sdk-s3/1.75.0
-      Content-Encoding:
-      - ASCII
-      Expect:
-      - 100-continue
-      Content-Md5:
-      - mgNkuembtIDdJeHwKEyFVQ==
-      Host:
-      - "<REDACTED_BUCKET_NAME>.s3.<REDACTED_REGION_NAME>.amazonaws.com"
-      X-Amz-Date:
-      - 20200813T135412Z
-      X-Amz-Content-Sha256:
-      - ed7002b439e9ac845f22357d822bac1444730fbdb6016d3ec9432297b9ec9f73
-      Authorization:
-      - AWS4-HMAC-SHA256 Credential=<REDACTED_ACCESS_KEY>/20200813/<REDACTED_REGION_NAME>/s3/aws4_request,
-        SignedHeaders=content-encoding;content-md5;host;user-agent;x-amz-content-sha256;x-amz-date,
-        Signature=82f622ee12ae96359f82c219a8ff98b319ae50627105d57e8a318604ed5ba610
-      Content-Length:
-      - '7'
-      Accept:
-      - "*/*"
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      X-Amz-Id-2:
-      - rH3ILCPD/GfkjJjSxo5yG1ar/R+iQHhw2kSqWHemWulLHyhvQEwISbtG5xGTUZOLSdu+45++aJ0=
-      X-Amz-Request-Id:
-      - 280A89561B09F6BD
-      Date:
-      - Thu, 13 Aug 2020 13:54:14 GMT
-      Etag:
-      - '"9a0364b9e99bb480dd25e1f0284c8555"'
-      Content-Length:
-      - '0'
-      Server:
-      - AmazonS3
-    body:
-      encoding: UTF-8
-      string: ''
-  recorded_at: Thu, 13 Aug 2020 13:54:14 GMT
-- request:
-    method: head
-    uri: https://<REDACTED_BUCKET_NAME>.s3.<REDACTED_REGION_NAME>.amazonaws.com/gemstash/s3_storage/TEST_S3_SPEC_FOLDER/private/gems/existing-f4e0ac58eb46d88efc451c164db3b837/properties.yaml
-    body:
-      encoding: UTF-8
-      string: ''
-    headers:
-      Content-Type:
-      - ''
-      Accept-Encoding:
-      - ''
-      User-Agent:
-      - aws-sdk-ruby3/3.104.3 ruby/2.6.3 x86_64-darwin19 aws-sdk-s3/1.75.0
-      Host:
-      - "<REDACTED_BUCKET_NAME>.s3.<REDACTED_REGION_NAME>.amazonaws.com"
-      X-Amz-Date:
-      - 20200813T135414Z
-      X-Amz-Content-Sha256:
-      - e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855
-      Authorization:
-      - AWS4-HMAC-SHA256 Credential=<REDACTED_ACCESS_KEY>/20200813/<REDACTED_REGION_NAME>/s3/aws4_request,
-        SignedHeaders=host;user-agent;x-amz-content-sha256;x-amz-date, Signature=7f7f8a2d04af934b08e877f786070707006290dd4940fac834de3d1de5f97ef5
-      Content-Length:
-      - '0'
-      Accept:
-      - "*/*"
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      X-Amz-Id-2:
-      - 1oHgHmAK/uQ147XZ7WB1VTCrTlgR15IwjWYGS5FOsRZ/vAP+QRvVen3tj6AukS3MeYTDzH5Nnmw=
-      X-Amz-Request-Id:
-      - 46274E6E4640428F
-      Date:
-      - Thu, 13 Aug 2020 13:54:15 GMT
-      Last-Modified:
-      - Thu, 13 Aug 2020 13:54:13 GMT
-      Etag:
-      - '"b79e501aa444a9e22294e4ad7eb54022"'
-      Content-Encoding:
-      - ASCII
-      Accept-Ranges:
-      - bytes
-      Content-Type:
-      - ''
-      Content-Length:
-      - '69'
-      Server:
-      - AmazonS3
-    body:
-      encoding: UTF-8
-      string: ''
-  recorded_at: Thu, 13 Aug 2020 13:54:15 GMT
-- request:
-    method: get
-    uri: https://<REDACTED_BUCKET_NAME>.s3.<REDACTED_REGION_NAME>.amazonaws.com/gemstash/s3_storage/TEST_S3_SPEC_FOLDER/private/gems/existing-f4e0ac58eb46d88efc451c164db3b837/properties.yaml
-    body:
-      encoding: UTF-8
-      string: ''
-    headers:
-      Content-Type:
-      - ''
-      Accept-Encoding:
-      - ''
-      User-Agent:
-      - aws-sdk-ruby3/3.104.3 ruby/2.6.3 x86_64-darwin19 aws-sdk-s3/1.75.0
-      Host:
-      - "<REDACTED_BUCKET_NAME>.s3.<REDACTED_REGION_NAME>.amazonaws.com"
-      X-Amz-Date:
-      - 20200813T135415Z
-      X-Amz-Content-Sha256:
-      - e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855
-      Authorization:
-      - AWS4-HMAC-SHA256 Credential=<REDACTED_ACCESS_KEY>/20200813/<REDACTED_REGION_NAME>/s3/aws4_request,
-        SignedHeaders=host;user-agent;x-amz-content-sha256;x-amz-date, Signature=81cc00d6323ca0a9cb698db8457c0f64becb786f8c50f16ae10ab42b2380ec24
-      Content-Length:
-      - '0'
-      Accept:
-      - "*/*"
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      X-Amz-Id-2:
-      - MPlPSmzRmnj4BzT64a+3DZiEkZOPeYPh6b56be4tzm83fVYajlXO7BUlXMgx3Fak3vJ4FltILBc=
-      X-Amz-Request-Id:
-      - 8Z1HDHDN3P6S6S9P
-      Date:
-      - Thu, 13 Aug 2020 13:54:16 GMT
-      Last-Modified:
-      - Thu, 13 Aug 2020 13:54:13 GMT
-      Etag:
-      - '"b79e501aa444a9e22294e4ad7eb54022"'
-      Content-Encoding:
-      - ASCII
-      Accept-Ranges:
-      - bytes
-      Content-Type:
-      - ''
-      Content-Length:
-      - '69'
-      Server:
-      - AmazonS3
-    body:
-      encoding: UTF-8
-      string: "---\n:gemstash_resource_version: 1\n:foo: one\n:bar: \n:baz:\n  :qux:
-        two\n"
-  recorded_at: Thu, 13 Aug 2020 13:54:15 GMT
-- request:
-    method: put
-    uri: https://<REDACTED_BUCKET_NAME>.s3.<REDACTED_REGION_NAME>.amazonaws.com/gemstash/s3_storage/TEST_S3_SPEC_FOLDER/private/gems/existing-f4e0ac58eb46d88efc451c164db3b837/properties.yaml
-    body:
-      encoding: UTF-8
-      string: "---\n:gemstash_resource_version: 1\n:foo: one\n:bar: \n:baz:\n  :qux:
-        two\n"
-    headers:
-      Content-Type:
-      - ''
-      Accept-Encoding:
-      - ''
-      User-Agent:
-      - aws-sdk-ruby3/3.104.3 ruby/2.6.3 x86_64-darwin19 aws-sdk-s3/1.75.0
-      Content-Encoding:
-      - ASCII
-      Expect:
-      - 100-continue
-      Content-Md5:
-      - t55QGqREqeIilOStfrVAIg==
-      Host:
-      - "<REDACTED_BUCKET_NAME>.s3.<REDACTED_REGION_NAME>.amazonaws.com"
-      X-Amz-Date:
-      - 20200813T135415Z
-      X-Amz-Content-Sha256:
-      - 9593548c4e0f8690e71e1143678921bf190edfeb41dfe64f1dd91ef62b2eddd7
-      Authorization:
-      - AWS4-HMAC-SHA256 Credential=<REDACTED_ACCESS_KEY>/20200813/<REDACTED_REGION_NAME>/s3/aws4_request,
-        SignedHeaders=content-encoding;content-md5;host;user-agent;x-amz-content-sha256;x-amz-date,
-        Signature=384ab0fb2e636190f6d9cee3dd51be063580d7949ece120c853077a42dcadee4
-      Content-Length:
-      - '69'
-      Accept:
-      - "*/*"
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      X-Amz-Id-2:
-      - ld/CMF5nOhuD8TmPV0fO2ZdhPoXSNqUG796BrScOB4ZG6Lzr7eUJOSnji9dkiM2/RGF65CLs+JQ=
-      X-Amz-Request-Id:
-      - 062C995FFCD66A2B
-      Date:
-      - Thu, 13 Aug 2020 13:54:17 GMT
-      Etag:
-      - '"b79e501aa444a9e22294e4ad7eb54022"'
-      Content-Length:
-      - '0'
-      Server:
-      - AmazonS3
-    body:
-      encoding: UTF-8
-      string: ''
-  recorded_at: Thu, 13 Aug 2020 13:54:16 GMT
-recorded_with: VCR 6.0.0
