--- fixtures/vcr_cassettes/Gemstash_S3/with_a_valid_storage/returns_a_non_existing_resource_when_requested.yml
@@ -1,50 +1,0 @@
----
-http_interactions:
-- request:
-    method: head
-    uri: https://<REDACTED_BUCKET_NAME>.s3.<REDACTED_REGION_NAME>.amazonaws.com/gemstash/s3_storage/TEST_S3_SPEC_FOLDER/private/gems/an_id-79453b648e93a9ef31ce8a13121eeca6/properties.yaml
-    body:
-      encoding: UTF-8
-      string: ''
-    headers:
-      Content-Type:
-      - ''
-      Accept-Encoding:
-      - ''
-      User-Agent:
-      - aws-sdk-ruby3/3.104.3 ruby/2.6.3 x86_64-darwin19 aws-sdk-s3/1.75.0
-      Host:
-      - "<REDACTED_BUCKET_NAME>.s3.<REDACTED_REGION_NAME>.amazonaws.com"
-      X-Amz-Date:
-      - 20200813T135014Z
-      X-Amz-Content-Sha256:
-      - e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855
-      Authorization:
-      - AWS4-HMAC-SHA256 Credential=<REDACTED_ACCESS_KEY>/20200813/<REDACTED_REGION_NAME>/s3/aws4_request,
-        SignedHeaders=host;user-agent;x-amz-content-sha256;x-amz-date, Signature=d7dc764dd54aeeaac0f3ae0d5910f2846467add73545826e2baed6ca805725c0
-      Content-Length:
-      - '0'
-      Accept:
-      - "*/*"
-  response:
-    status:
-      code: 404
-      message: Not Found
-    headers:
-      X-Amz-Request-Id:
-      - EF97DA0166544A57
-      X-Amz-Id-2:
-      - KQDxXmiZRvVSuow2EIjdeHrWgGfn3WkqPeymmiYmsD7MRNGahLcCluU8Hjklglhsn+NZ+B7NVfc=
-      Content-Type:
-      - application/xml
-      Transfer-Encoding:
-      - chunked
-      Date:
-      - Thu, 13 Aug 2020 13:50:15 GMT
-      Server:
-      - AmazonS3
-    body:
-      encoding: UTF-8
-      string: ''
-  recorded_at: Thu, 13 Aug 2020 13:50:15 GMT
-recorded_with: VCR 6.0.0
