+++ lib/gemstash/specs_builder.rb
@@ -19,7 +19,7 @@ module Gemstash
     end
 
     def self.invalidate_stored
-      storage = Gemstash::LocalStorage.for("private").for("specs_collection")
+      storage = Gemstash::Storage.for("private").for("specs_collection")
       storage.resource("specs.4.8.gz").delete(:specs)
       storage.resource("latest_specs.4.8.gz").delete(:specs)
       storage.resource("prerelease_specs.4.8.gz").delete(:specs)
@@ -46,7 +46,7 @@ module Gemstash
   private
 
     def storage
-      @storage ||= Gemstash::LocalStorage.for("private").for("specs_collection")
+      @storage ||= Gemstash::Storage.for("private").for("specs_collection")
     end
 
     def fetch_resource
