+++ spec/gemstash/gem_pusher_spec.rb
@@ -10,7 +10,7 @@ RSpec.describe Gemstash::GemPusher do
   let(:auth_key) { "auth-key" }
   let(:invalid_auth_key) { "invalid-auth-key" }
   let(:auth_key_without_permission) { "auth-key-without-permission" }
-  let(:storage) { Gemstash::LocalStorage.for("private").for("gems") }
+  let(:storage) { Gemstash::Storage.for("private").for("gems") }
 
   before do
     Gemstash::Authorization.authorize(auth_key, "all")
