+++ spec/gemstash/gem_yanker_spec.rb
@@ -9,7 +9,7 @@ RSpec.describe Gemstash::GemYanker do
   let(:auth_key) { "auth-key" }
   let(:invalid_auth_key) { "invalid-auth-key" }
   let(:auth_key_without_permission) { "auth-key-without-permission" }
-  let(:storage) { Gemstash::LocalStorage.for("private").for("gems") }
+  let(:storage) { Gemstash::Storage.for("private").for("gems") }
   let(:deps) { Gemstash::Dependencies.for_private }
   let(:gem_name) { "example" }
   let(:gem_version) { "0.1.0" }
